refactor: CommentAddDialog
